uno sayti/requirements.txt
aiohttp==3.9.3
python-socketio[aiohttp]==5.11.1
gunicorn==21.2.0
